+++ framework/model/src/main/java/is/codion/framework/model/package-info.java
@@ -22,7 +22,7 @@
  * <li>{@link is.codion.framework.model.EntityApplicationModel#USER_PREFERENCES}
  * <li>{@link is.codion.framework.model.EntityApplicationModel#PREFERENCES_KEY}
  * <li>{@link is.codion.framework.model.EntityEditModel#EDIT_EVENTS}
- * <li>{@link is.codion.framework.model.EntityEditModel.EntityEditor#PERSIST_FOREIGN_KEYS}
+ * <li>{@link is.codion.framework.model.EntityEditor#PERSIST_FOREIGN_KEYS}
  * <li>{@link is.codion.framework.model.EntityQueryModel#LIMIT}
  * <li>{@link is.codion.framework.model.EntityTableModel#ON_INSERT}
  * <li>{@link is.codion.framework.model.EntityTableModel#ORDER_QUERY}