+++ changelog.md
@@ -11,6 +11,7 @@ Codion Change Log
 - EntityEditor.EditorModels added, related refactoring.
 - EntityEditor.SearchModels added.
 - EntityEditor.replace() added, DefaultEntityEditor.setOrDefaults() now private.
+- EntityEditModel.EntityEditor moved up to package level, EntityEditModel.EditorValue moved to EntityEditor.
 ### is.codion.swing.framework.model
 - SwingEntityEditModel.createComboBoxModel() memory leak bug fixed, refresh listener now only handles managed column combo box models, related refactoring.
 - SwingEntityEditor added along with SwingEditorModels.